[ minor ] commenting code, fixing a test case
Also one optimisation: no need to process the whole file if we're
already beyond the range of interest. I was hoping we'd be able to
seek directly to the start of the range but unfortunately it doesn't
seem possible to seek by row + col, only by offset in the file.